feat: configure frontend to work with both languages
