+++ .env.sample
@@ -14,7 +14,5 @@ ADMIN_EMAIL=admin@example.com
 ADMIN_PASSWORD= password
 BAMBOOHR_SUBDOMAIN=<Ask a team mate>
 BAMBOOHR_API_KEY=<Ask a team mate>
-RAILS_BACKEND_URI=http://localhost:5001
-PHOENIX_BACKEND_URI=http://localhost:4000
-WEBSOCKET_URI=ws://localhost:4000
-BACKEND_LANGUAGE=elixir
+BACKEND_URL=http://localhost:5001
+BACKEND_LANGUAGE=ruby
